scripts/requirements.txt
GitPython
twine==2.0.0
setuptools